% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cultivo2009.R
\docType{data}
\name{cultivo2009}
\alias{cultivo2009}
\title{Data for Genotypes by Environment Interaction (GEI)}
\format{A \code{data.frame} 1320 obs. of  6 variables.}
\usage{
data(cultivo2009)
}
\description{
\code{cultivo2009} is used for performing Genotypes by Environment Interaction (GEI) Analysis.
}
\details{
\itemize{
       \item{Gen} Genotype
       \item{Institute} Institute
       \item{Rep} Replicate
       \item{Block} Block
       \item{Env} Environment
       \item{Yield} Yield Response
       }
}
\examples{
data(cultivo2009)

}
\references{
Perez-Elizalde, S., Jarquin, D., and Crossa, J. (2011)
 A General Bayesian Estimation Method of Linear–Bilinear Models
 Applied to Plant Breeding Trials With Genotype × Environment Interaction.
 \emph{Journal of Agricultural, Biological, and Environmental Statistics},
  17, 15–37.  (\href{https://link.springer.com/article/10.1007/s13253-011-0063-9}{doi:10.1007/s13253-011-0063-9})
}
\author{
\enumerate{
    \item Muhammad Yaseen (\email{myaseen208@gmail.com})
    \item Diego Jarquin (\email{diego.jarquin@gmail.com})
    \item Sergio Perez-Elizalde (\email{sergiop@colpos.mx})
    \item Juan Burgueño (\email{j.burgueno@cgiar.org})
    \item Jose Crossa (\email{j.crossa@cgiar.org})
   }
}
\keyword{datasets}
