% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_k.R
\name{matrix_k}
\alias{matrix_k}
\alias{matrix_k.default}
\title{k Matrix}
\usage{
matrix_k(n)

\method{matrix_k}{default}(n)
}
\arguments{
\item{n}{Number of columns}
}
\value{
Matrix
}
\description{
Gives k matrix
}
\references{
Perez-Elizalde, S., Jarquin, D., and Crossa, J. (2011)
 A General Bayesian Estimation Method of Linear–Bilinear Models
 Applied to Plant Breeding Trials With Genotype × Environment Interaction.
 \emph{Journal of Agricultural, Biological, and Environmental Statistics},
  17, 15–37.  (\href{https://link.springer.com/article/10.1007/s13253-011-0063-9}{doi:10.1007/s13253-011-0063-9})
}
\author{
\enumerate{
    \item Muhammad Yaseen (\email{myaseen208@gmail.com})
   }
}
