% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthnorm.R
\name{orthnorm}
\alias{orthnorm}
\alias{orthnorm.default}
\title{Orthogonal Normalization}
\usage{
orthnorm(u = NULL, basis = TRUE, norm = TRUE)

\method{orthnorm}{default}(u = NULL, basis = TRUE, norm = TRUE)
}
\arguments{
\item{u}{Matrix}

\item{basis}{Logical argument by default TRUE}

\item{norm}{Logical argument by default TRUE}
}
\value{
Matrix
}
\description{
Perform Orthogonal Normalization of a matrix
}
\references{
Perez-Elizalde, S., Jarquin, D., and Crossa, J. (2011)
 A General Bayesian Estimation Method of Linear–Bilinear Models
 Applied to Plant Breeding Trials With Genotype × Environment Interaction.
 \emph{Journal of Agricultural, Biological, and Environmental Statistics},
  17, 15–37.  (\href{https://link.springer.com/article/10.1007/s13253-011-0063-9}{doi:10.1007/s13253-011-0063-9})
}
\author{
\enumerate{
    \item Muhammad Yaseen (\email{myaseen208@gmail.com})
   }
}
